adds endpoint testing
