+++ client/src/app.jsx
@@ -1,20 +1,34 @@
-import React from 'react';
+// import React from 'react';
+import React, { useState, useEffect } from 'react';
 import ReactDOM from 'react-dom';
-// const axios = require('axios');
+import styled from 'styled-components'
+import axios from 'axios';
 
-class App extends React.Component {
-  constructor(props) {
-    super(props)
-    this.state = {
-      test: 'test'
-    };
-  };
-  render () {
-    return (
-      <div>Hello</div>
-    )
-  }
-};
+
+const App = () => {
+  const [name, setName] = useState('tom');
+
+  return (
+    <div>
+      <p>hey hey hey</p>
+      <p onClick={() => setName('jack')}>{name}</p>
+    </div>
+  )
+}
+
+// class App extends React.Component {
+//   constructor(props) {
+//     super(props)
+//     this.state = {
+//       test: 'test'
+//     };
+//   };
+//   render () {
+//     return (
+//       <div className="container">hola</div>
+//     )
+//   }
+// };
