+++ server/start.js
@@ -1,0 +1,2 @@
+const app = require("./server.js");
+app.listen(3001);
